+++ src/map/utils/petutils.cpp
@@ -1252,10 +1252,10 @@ namespace petutils
         PPet->UpdateHealth();
 
         // boost stats by 10%
-        PPet->addModifier(Mod::ATTP, (int32)(rate * 100.0f));
-        PPet->addModifier(Mod::ACC, (int32)(rate * 100.0f));
-        PPet->addModifier(Mod::EVA, (int32)(rate * 100.0f));
-        PPet->addModifier(Mod::DEFP, (int32)(rate * 100.0f));
+        PPet->addModifier(Mod::ATTP, (int16)(rate * 100.0f));
+        PPet->addModifier(Mod::ACC, (int16)(rate * 100.0f));
+        PPet->addModifier(Mod::EVA, (int16)(rate * 100.0f));
+        PPet->addModifier(Mod::DEFP, (int16)(rate * 100.0f));
     }
 
     void LoadPet(CBattleEntity* PMaster, uint32 PetID, bool spawningFromZone)
